Fixed a bug that blocked the move of logs when full.
git-svn-id: http://deveeldb.googlecode.com/svn/trunk@136 de0d096c-ad4d-11de-a70b-4539db072d2e

+++ deveeldb/Deveel.Diagnostics/LogWriter.cs
@@ -103,7 +103,8 @@ namespace Deveel.Diagnostics {
 				for (int i = archive_count - 1; i > 0; --i) {
 					string source = log_file + "." + i;
 					string dest = log_file + "." + (i + 1);
-					File.Move(source, dest);
+					if (File.Exists(source))
+						File.Move(source, dest);
 				}
 				File.Move(log_file, log_file + ".1");
